ENH: Fix problem if two NIRFast in the path
This helps address the problem of someone having 2 versions of nirfast
in the path.  This is the only place I’ve run into a problem with that.
This solution assumes that the Nirfast version higher in the path list
is what the user wants to use. This is required for the “Select System”
dropdown list to be populated, which is critical for making final
placements of the source-detector pairs.

Authored-by: Scott C. Davis <Scott.C.Davis@dartmouth.edu>